[FE] refactor: - create func component - Header; - rename directory;
