+++ src/components/offer-card/OfferCard.jsx
