+++ src/components/suggestion-list/SuggestionList.jsx
@@ -1,6 +1,7 @@
 import React from 'react';
 import PropTypes from 'prop-types';
-import OfferCard from "../offerCard/OfferCard";
+import OfferCard from "../offer-card/OfferCard";
+import Header from "../header/Header";
 
 class SuggestionList extends React.Component {
   constructor(props) {
@@ -22,28 +23,7 @@ class SuggestionList extends React.Component {
     const {cards} = this.props;
     return (
       <div className="page page--gray page--main">
-        <header className="header">
-          <div className="container">
-            <div className="header__wrapper">
-              <div className="header__left">
-                <a className="header__logo-link header__logo-link--active">
-                  <img className="header__logo" src="img/logo.svg" alt="6 cities logo" width="81" height="41"/>
-                </a>
-              </div>
-              <nav className="header__nav">
-                <ul className="header__nav-list">
-                  <li className="header__nav-item user">
-                    <a className="header__nav-link header__nav-link--profile" href="#">
-                      <div className="header__avatar-wrapper user__avatar-wrapper">
-                      </div>
-                      <span className="header__user-name user__name">Oliver.conner@gmail.com</span>
-                    </a>
-                  </li>
-                </ul>
-              </nav>
-            </div>
-          </div>
-        </header>
+        <Header isMainPage={true}/>
 
         <main className="page__main page__main--index">
           <h1 className="visually-hidden">Cities</h1>
